+++ sources/src/api/noise/Noise.cpp
@@ -5,7 +5,7 @@
 #include "inc/api/noise/Noise.h"
 #include <utility>
 
-Noise::Noise(BitMap<int> bitmap, const uint seed, const uint width, const uint height, const uint frequency)
+Noise::Noise(BitMap<int> bitmap, const size_t seed, const size_t width, const size_t height, const size_t frequency)
         : m_bitmap(std::move(bitmap)), m_seed(seed), m_width(width), m_height(height), m_frequency(frequency) {}
 
 const BitMap<int>& Noise::getBitmap() const {
